+++ src/main/java/lotto/domain/dto/Rate.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.dto;
 
 public class Rate {
     private final int rate;
