+++ src/main/java/lotto/domain/dto/TotalPrice.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.dto;
 
 public class TotalPrice {
     private final long price;
